+++ integration/testing-tools/src/contentsOf.ts
@@ -1,0 +1,6 @@
+import fs = require('fs');
+import path = require('path');
+
+export function contentsOf(...pathSegments: string[]) {
+    return fs.readFileSync(path.join(...pathSegments), 'utf8');
+}