+++ packages/serenity-bdd/src/stage/crew/serenity-bdd-reporter/reports/SceneReportId.ts
@@ -1,5 +1,8 @@
 import { ensure, isDefined, isGreaterThan, property, TinyType } from 'tiny-types';
 
+/**
+ * @package
+ */
 export class SceneReportId extends TinyType {
 
     public readonly values: string[] = [];